Java: fix specs because now Java integers without default case are considered nullable, and thus are boxed

+++ spec/java/src/io/kaitai/struct/spec/TestSwitchIntegers2.java
@@ -11,7 +11,7 @@ public class TestSwitchIntegers2 extends CommonSpec {
         SwitchIntegers2 r = SwitchIntegers2.fromFile(SRC_DIR + "switch_integers.bin");
 
         assertEquals(r.code(), 1);
-        assertEquals(r.len(), 7);
+        assertEquals(r.len().longValue(), 7);
         assertEquals(r.ham(), new byte[] { 2, 64, 64, 4, 55, 19, 0 });
         assertEquals(r.padding().intValue(), 0);
         assertEquals(r.lenModStr(), "13");
